+++ src/lib/components/CardBoard.svelte
@@ -1,24 +1,10 @@
 <script lang="ts">
-  import { createEventDispatcher, onMount, type ComponentProps } from 'svelte';
+  import { createEventDispatcher, type ComponentProps } from 'svelte';
   import type { Keyed } from '$lib/types';
   import Card from './Card.svelte';
+  import { fullscreen } from '$lib/actions';
 
   export let cards: Keyed<ComponentProps<Card>>[];
-
-  let viewport: HTMLDivElement;
-
-  onMount(() => {
-    const resizeViewport = () => {
-      viewport.style.height = `${window.innerHeight}px`;
-    };
-
-    resizeViewport();
-    window.addEventListener('resize', resizeViewport);
-
-    return () => {
-      window.removeEventListener('resize', resizeViewport);
-    };
-  });
 
   const dispatch = createEventDispatcher<{ cardSelected: { card: ComponentProps<Card> } }>();
   const propagateSelection = (data: CustomEvent<{ name: string }>) => {
@@ -32,8 +18,8 @@
 </script>
 
 <div
+  use:fullscreen
   class="viewport bg-base-100"
-  bind:this={viewport}
   >
   <div class="grid gap-2 lg:grid-cols-3 md:grid-cols-2 sm:grid-cols-1 overscroll-auto">
     {#each cards as { id, ...cardProps } (id)}
